feat(db): 彻底迁移 PostgreSQL——删除全部 SQLite 支持 + 测试迁移至 PG
- 删 modernc/sqlite 依赖、openSQLite、DriverSQLite、SQLite 迁移目录(39个)
- db.go: PG-only(连接池 200/50 + TimeZone=Asia/Shanghai + rewrite 层保留
  承载 ? →  转译,SQL 无需全改)
- cmd: -db-driver 仅 pg/pg-external;删除 cmd/migrate-sqlite-pg 与 Dockerfile 构建项
- migrate.go: 仅读 migrations-pg(30个,与 SQLite 语义对齐)
- sysinfo/collectDBStats: PG 固定
- 生产修复: instr→POSITION(lower/usrname)、LastInsertId→RETURNING id、
  quota_tokens/prompt_tokens/completion_tokens/cache_prompt_tokens → BIGINT
- 测试: 每测试独立临时 PG 库(NewTestDB/pgx,CREATE/DROP DATABASE),
  17 个测试文件从 DBConfig{Path} 迁移;sqlite_master→pg_indexes/pg_tables;
  datetime('now')→now()-interval;LastInsertId→RETURNING;COLLATE NOCASE→LOWER
- 全部 Go 测试绿(10 包: agentshare/bootstrap/capabilities/llmgateway/
  marketplace/serverauth/serverstore/sharedskills/util/channels)

+++ server/internal/serverstore/migrations-pg/0030_usage_cache_tokens.sql
@@ -1,4 +1,4 @@
 -- 0030: usage 增加缓存命中输入 token 数(DeepSeek 缓存计费)。
 -- prompt_cache_hit_tokens = 上游返回的缓存命中输入 token;未配置时按 0 计,
 -- 命中部分按 models.cache_input_price_per_1m(0029)计费,未配置则回退输入价。
-ALTER TABLE usage ADD COLUMN cache_prompt_tokens INTEGER NOT NULL DEFAULT 0;
+ALTER TABLE usage ADD COLUMN cache_prompt_tokens BIGINT NOT NULL DEFAULT 0;
